pyproject.toml
[tool.poetry]
name = "fastapi-kafka"
version = "0.1.0"
description = ""
authors = ["Ahmed Nafies <ahmed.nafies@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_kafka"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
